Struct strymon_executor::executable::Builder

pub struct Builder { /* fields omitted */ }

A builder for configuring a new job process.

In order to spawn a configured binary, it has to be passed to a running ProcessService instance:

Examples

use strymon_executor::executable::Builder;

let mut process = Builder::new("my_job", &["arg1"]).unwrap();
process
    .process(0)
    .threads(4);

Methods

impl Builder

pub fn new<T, S, I>(executable: T, args: I) -> Result<Self, SpawnError> where
T: AsRef<Path>,
S: AsRef<OsStr>,
I: IntoIterator<Item = S>,

Creates a new builder for the given native executable file.

pub fn threads(&mut self, threads: usize) -> &mut Self

Sets the number of Timely threads per worker (default: 1)

pub fn process(&mut self, process: usize) -> &mut Self

Sets the current Timely process id (default: 0)

pub fn hostlist(&mut self, hostlist: Vec<String>) -> &mut Self

Specify the host names of all Timely processes (default: [])

pub fn working_directory<P: AsRef<Path>>(&mut self, workdir: P) -> &mut Self

Sets the working directory of the child process (default: std::env::current_dir())