Struct strymon_executor::executable::ProcessService

pub struct ProcessService { /* fields omitted */ }

Strymon job process spawning and supervision service.

This structure is responsible for keeping track of the currently running child processes, their output and handles potential termination requests.

All supervised children are terminated when the tokio event loop is dropped, though this might change in the future.

Methods

impl ProcessService

pub fn new(handle: &Handle, coord: String, hostname: String) -> Self

Create a new service instance.

This will spawn processes using the tokio event loop of the handle. The coord and hostname configuration values are passed down to the spawned jobs.

pub fn spawn(&mut self, id: JobId, builder: Builder) -> Result<(), SpawnError>

Spawns a new process to be supervised by this process service.

The Timely configuration stored in the Builder is extracted and passed down to the spawned child process.

pub fn terminate(&mut self, job_id: JobId) -> Result<(), TerminateError>

Tries to terminate the process assigned to a given job identifier.

This will result in a SIGTERM being send to the child on Unix platforms.