Struct strymon_executor::Builder

pub struct Builder { /* fields omitted */ }

A builder type to initialize and run a new executor instance.

Examples

use strymon_executor::Builder;

let mut executor = Builder::default();
executor
    .hostname("worker7".to_string())
    .coordinator("master:9189".to_string());
executor.start().expect("failed to start executor");

Methods

impl Builder

pub fn hostname(&mut self, host: String) -> &mut Self

Sets the externally reachable hostname of this machine (default: inferred).

pub fn coordinator(&mut self, coord: String) -> &mut Self

Sets the address of the coordinator (default: "localhost:9189")

pub fn ports(&mut self, min: u16, max: u16) -> &mut Self

Sets a range of ports to be allocated to timely_communication workers (default: 2101-4101)

pub fn workdir<P: AsRef<OsStr>>(&mut self, workdir: &P) -> &mut Self

Sets the base path for job working directories (default: "jobs").

impl Builder

pub fn start(self) -> Result<(), Error>

Starts a new executor instance and blocks until termination.

Initiates the connection to the coordinator, sends an executor registration request and upon successful initialization, drives the ExecutorService with incoming requests from the coordinator.

The service is terminated either until SIGTERM is received or the connection to the coordinator is closed.