Struct strymon_executor::ExecutorService

pub struct ExecutorService { /* fields omitted */ }

A running Strymon executor instance.

Methods

impl ExecutorService

pub fn new(
id: ExecutorId,
coord: String,
workdir: PathBuf,
network: Network,
handle: Handle
) -> Self

Creates a new instance for an newly registered executor.

This assumes that a AddExecutor request has been sent to the coordinator. This constructor takes the following arguments:

id: The identifier assigned to us in the AddExecutor response from the coordinator.

coord: The URI of the coordinator which will be passed down to spawned jobs.

workdir: The base path used to create working directories for spawned jobs.

handle: A tokio-core reactor handle, used to create the ProcessService.

pub fn dispatch(&mut self, req: RequestBuf<ExecutorRPC>) -> Result<(), Error>

Decode and handle a request sent by the coordinator.